Rename model folder to models Resolves #28
