+++ user/usecase/user.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/bashmohandes/go-askme/answer"
-	"github.com/bashmohandes/go-askme/model"
+	"github.com/bashmohandes/go-askme/models"
 	"github.com/bashmohandes/go-askme/question"
 	"github.com/bashmohandes/go-askme/user"
 )
